Ensure Watchlist is deleted when associated user is removed

+++ src/Persistence/Repositories/UserRepository.cs
@@ -27,4 +27,13 @@ public class UserRepository(AppDbContext context) : IUserRepository
     public void AddUser(User user) => context.Users.Add(user);
 
     public void RemoveUser(User user) => context.Users.Remove(user);
+
+    public async Task RemoveUserWatchlistAsync(string userId)
+    {
+        var watchlist = await context.Watchlist
+            .Where(wl => wl.UserId == userId)
+            .ToListAsync();
+
+        context.Watchlist.RemoveRange(watchlist);
+    }
 }
